# nodeapp/swagger.yaml
swagger: '2.0'
info:
  description: 'Api de NodeAPP'
  version: '1.1.0'
  title: 'NodeApp'
  termsOfService: 'http://swagger.io/terms/'
  contact:
    email: 'apiteam@swagger.io'
  license:
    name: 'Apache 2.0'
    url: 'http://www.apache.org/licenses/LICENSE-2.0.html'
host: 'petstore.swagger.io'
basePath: '/v2'
...
schemes:
  - 'https'
  - 'http'
paths:
  /agentes:
    post:
      tags:
        - 'pet'
      summary: 'Add a new agente to the store'
      description: ''
      operationId: 'addPet'
      consumes:
        - 'application/json'
        - 'application/xml'
      produces:
        - 'application/xml'
        - 'application/json'
      parameters:
        - in: 'body'
          name: 'body'
          description: 'Pet object that needs to be added to the store'
          required: true
          schema:
            $ref: '#/definitions/Pet'
      responses:
        '405':
          description: 'Invalid input'
      security:
        - petstore_auth:
            - 'write:pets'
            - 'read:pets'
    put:
      tags:
        - 'pet'
      summary: 'Update an existing pet'
      description: ''
      operationId: 'updatePet'
      consumes:
        - 'application/json'
        - 'application/xml'
      produces:
        - 'application/xml'
        - 'application/json'
      parameters:
        - in: 'body'
          name: 'body'
          description: 'Pet object that needs to be added to the store'
          required: true
          schema:
            $ref: '#/definitions/Pet'
      responses:
        '400':
          description: 'Invalid ID supplied'
        '404':
          description: 'Pet not found'
        '405':
          description: 'Validation exception'
      security:
        - petstore_auth:
            - 'write:pets'
            - 'read:pets'
...
securityDefinitions:
  petstore_auth:
    type: 'oauth2'
    authorizationUrl: 'http://petstore.swagger.io/oauth/dialog'
    flow: 'implicit'
    scopes:
      write:pets: 'modify pets in your account'
      read:pets: 'read your pets'
  api_key:
    type: 'apiKey'
    name: 'api_key'
    in: 'header'
definitions:
...
  Category:
    type: 'object'
    properties:
      id:
        type: 'integer'
        format: 'int64'
      name:
        type: 'string'
    xml:
      name: 'Category'
...
  Tag:
    type: 'object'
    properties:
      id:
        type: 'integer'
        format: 'int64'
      name:
        type: 'string'
    xml:
      name: 'Tag'
  Pet:
    type: 'object'
    required:
      - 'name'
      - 'photoUrls'
    properties:
      id:
        type: 'integer'
        format: 'int64'
      category:
        $ref: '#/definitions/Category'
      name:
        type: 'string'
        example: 'doggie'
      photoUrls:
        type: 'array'
        xml:
          name: 'photoUrl'
          wrapped: true
        items:
          type: 'string'
      tags:
        type: 'array'
        xml:
          name: 'tag'
          wrapped: true
        items:
          $ref: '#/definitions/Tag'
      status:
        type: 'string'
        description: 'pet status in the store'
        enum:
          - 'available'
          - 'pending'
          - 'sold'
    xml:
      name: 'Pet'
...